# Conflicts:
#	source/jni/api/VFrameSmooth.cpp
#	source/jni/gui/VRMenuObjectLocal.cpp

+++ source/jni/api/VRotationState.h
@@ -7,7 +7,7 @@ NV_NAMESPACE_BEGIN
 struct VRotationState : public VQuatf
 {
     V3Vectf gyro;
-    longlong timestamp;
+    double timestamp;
 };
 
 NV_NAMESPACE_END
